WIP major refactor of Cluster and Shell-related classes.
Cluster now inherits from BaseShell, which is dynamically mixed-in with one of two classes: SshShell or LocalShell, depending on use.  BaseShell contains basic and shared functionality and the two mixins contain use-specific functionality and code for method- and attribute-access that needs to be handled differently in the two cases. In both cases, self.environ is a PseudoEnviron object that quacks like a dictionary but contains various extra/useful os.environ-related functionality.